Dump usage when we die.
I haven't used this for a while, and I remember what the usage
instructions are, nor could I quickly figure out how to dump them
to the console.

This has the drawback that the error message is now "hidden" at the
top of the (long) usage instructions, but I think it is more important
to output the usage.