Test the shmig script in repo against all three backends
If you are hacking on shmig, this is a much better model, as you
don't have to wait for the container to build to test your
changes.

Also, I frequently had inconsistent results running a docker
container with shmig.  I think docker does not like being started
and stopped frequenly in a loop.  It seemed to improve as docker
"warmed up", but it was still a pain.