Unbreak build
It was the quoting around "$args" that broke sqlite3 and mysql.

While here, implement SC2155 because ignoring errors is never good.

This also rolls back a bunch of changes from @shamrin's pull request;
for example, replacing "args="-U $LOGIN    $args" with
"args+=(-U "$LOGIN")".  That was not recommended by shellcheck so
I'll put that in another commit.